今日のぶん

+++ lib/log_getter.py
@@ -19,30 +19,35 @@ def create_ruru_id(ruru_number):
     return ruru_log_id
 
 
-def download_ruru_log(ruru_number):
+def download_ruru_log(ruru_number, save_path='../log'):
     """
+    指定されたるる鯖のナンバーのログを取得する。指定した場所にログがある場合は無を返す
     :param ruru_number: int
-    :return:
+    :param save_path: str
+    :return: bool
     """
-    save_path = '../log'
 
     # ファイルが無い場合は作成する
     if not os.path.exists(save_path):
         os.mkdir(save_path)
 
     file_pass = '../log/%s.html' % ruru_number  # 保存先のパス名
-    log_number = (create_ruru_id(ruru_number))  # ログのID
+    if os.path.isfile(file_pass):
+        print('EXIST %s' % file_pass)
+        return False
 
+    log_number = (create_ruru_id(ruru_number))  # ログのID
     ruru_url = 'https://ruru-jinro.net/%s/log%d.html' % (log_number, ruru_number)
     try:
         urllib.request.urlretrieve(ruru_url, file_pass)
         print('GET %s' % ruru_url)
+        return True
     except:
         print('ERROR %s' % ruru_url)
-        pass
+        return True
 
 if __name__ == '__main__':
-    for v in range(102212, 420000):
-        download_ruru_log(v)
-        time.sleep(1)
+    for v in range(186834, 420000):
+        if download_ruru_log(v):
+            time.sleep(1)
